+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Change categories:
 - Stabilised `decode_get_event_receiver_resp()`
 - Improved documentation of `struct pldm_package_format_pin`
 - base: Allow PLDM File Transfer to use MultiPartReceive decode API
+- Add feature flag to pldm package decoding APIs
 
 ### Deprecated